+++ controllers/trackController.js
@@ -1,6 +1,5 @@
 const bodyParser = require('body-parser');
 const Track = require('../models/trackSchema');
-const CartItem = require('../models/userCartSchema');
 const urlencodedParser = bodyParser.urlencoded({extended: false});
 
 module.exports = function (app) {
@@ -54,37 +53,5 @@ module.exports = function (app) {
                 res.render('track-add', {tracks: data, title: 'Track', username: ''});
             }
         });
-    });
-    app.post('/cart', urlencodedParser, function (req, res) {
-        var newCartItem = CartItem(req.body).save(function (err, data) {
-            if (err) throw err;
-            if (req.query.username != null) {
-                res.render('cart', {cartitems: data, title: 'Shopping Cart', username: req.query.username});
-            } else {
-                res.render('cart', {cartitems: data, title: 'Shopping Cart', username: ''});
-            }
-        });
-    });
-    app.get('/cart', function (req, res) {
-        CartItem.find({username: req.query.username}, function (err, data) {
-            if (err) throw err;
-            if (req.query.username != null) {
-                //console.log(data);
-                res.render('cart', {cartitems: data, title: 'Shopping Cart', username: req.query.username});
-            } else {
-                res.render('cart', {cartitems: data, title: 'Shopping Cart', username: ''});
-            }
-        });
-    });
-    app.delete('/cart/:_id', function (req, res) {
-        CartItem.find({_id: req.params._id})
-            .remove(function (err, data) {
-                if (err) throw err;
-                if (req.query.username != null) {
-                    res.render('cart', {cartitems: data, title: 'Shopping Cart', username: req.query.username});
-                } else {
-                    res.render('cart', {cartitems: data, title: 'Shopping Cart', username: ''});
-                }
-            });
     });
 };
